requirements.txt
discord.py==1.7.3
neuralintents==0.0.4
